Cleanup and refactoring (package organisation).
Also added needed extra /openbis to the OpenBIS URL
