+++ src/main/java/org/fairdom/openseekapi/facility/JSONCreator.java
@@ -1,4 +1,4 @@
-package org.fairdom;
+package org.fairdom.openseekapi.facility;
 
 import java.util.ArrayList;
 import java.util.HashMap;
