+++ src/main/java/org/fairdom/openseekapi/general/Authentication.java
@@ -1,4 +1,6 @@
-package org.fairdom;
+package org.fairdom.openseekapi.general;
+
+import org.fairdom.openseekapi.facility.SslCertificateHelper;
 
 import ch.ethz.sis.openbis.generic.asapi.v3.IApplicationServerApi;
 import ch.systemsx.cisd.common.spring.HttpInvokerUtils;
