+++ src/main/java/org/fairdom/openseekapi/general/AuthenticationException.java
@@ -1,4 +1,4 @@
-package org.fairdom;
+package org.fairdom.openseekapi.general;
 
 /**
  * @author Stuart Owen
