+++ src/test/java/org/fairdom/openseekapi/AuthenticationTest.java
@@ -1,7 +1,9 @@
-package org.fairdom;
+package org.fairdom.openseekapi;
 
 import static org.junit.Assert.assertTrue;
 
+import org.fairdom.openseekapi.general.Authentication;
+import org.fairdom.openseekapi.general.AuthenticationException;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.remoting.RemoteAccessException;
