+++ src/test/java/org/fairdom/openseekapi/Commons.java
@@ -1,7 +1,7 @@
-package org.fairdom;
+package org.fairdom.openseekapi;
 
 public class Commons {
-	public static final String TEST_OPENBIS_URL = "https://192.168.178.40/openbis/"; // "https://127.0.0.1:8443/openbis/openbis";
+	public static final String TEST_OPENBIS_URL = "https://192.168.178.40/openbis/openbis/"; // "https://127.0.0.1:8443/openbis/openbis";
 	public static final String TEST_OPENBIS_USER = "seek";
 	public static final String TEST_OPENBIS_PASSWORD = "seek";
 	public static final String TEST_OPENBIS_DATASTORE_ENDPOINT = "https://192.168.178.40/datastore_server/"; //"https://openbis-api.fair-dom.org/datastore_server";
