+++ src/test/java/org/fairdom/openseekapi/JSONHelper.java
@@ -1,4 +1,4 @@
-package org.fairdom;
+package org.fairdom.openseekapi;
 
 import static org.junit.Assert.assertTrue;
